chore: remove epoch card
