--- src/features/Overview/EpochCard/epochCard.module.css
@@ -1,15 +1,0 @@
-.stat-row {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-  width: 100%;
-
-  > div {
-    flex: 1 1 auto;
-    min-width: 180px;
-  }
-}
-
-.progress {
-  min-width: 140px;
-}
