+++ src/features/Overview/StatusCard/statusCard.module.css
@@ -9,3 +9,7 @@
     min-width: 180px;
   }
 }
+
+.progress {
+  min-width: 140px;
+}
